Add: new faker (#1324)
* Update: new example faker

* Update: 设置 Polar 图 label_opts 的 is_show 默认值为 False
