+++ example/bar_example.py
@@ -1,7 +1,7 @@
-from example.commons import Collector, Faker
 from pyecharts import options as opts
 from pyecharts.charts import Bar, Page
 from pyecharts.commons.utils import JsCode
+from pyecharts.faker import Collector, Faker
 from pyecharts.globals import ThemeType
 
 C = Collector()
