+++ example/boxplot_example.py
@@ -1,6 +1,6 @@
-from example.commons import Collector
 from pyecharts import options as opts
 from pyecharts.charts import Boxplot, Page
+from pyecharts.faker import Collector
 
 C = Collector()
