+++ example/graph_example.py
@@ -1,9 +1,9 @@
 import json
 import os
 
-from example.commons import Collector
 from pyecharts import options as opts
 from pyecharts.charts import Graph, Page
+from pyecharts.faker import Collector
 
 C = Collector()
