+++ example/line3d_example.py
@@ -1,8 +1,8 @@
 import math
 
-from example.commons import Collector, Faker
 from pyecharts import options as opts
 from pyecharts.charts import Line3D, Page
+from pyecharts.faker import Collector, Faker
 
 C = Collector()
