+++ example/scatter3d_example.py
@@ -1,8 +1,8 @@
 import random
 
-from example.commons import Collector, Faker
 from pyecharts import options as opts
 from pyecharts.charts import Page, Scatter3D
+from pyecharts.faker import Collector, Faker
 
 C = Collector()
