+++ example/treemap_example.py
@@ -1,9 +1,9 @@
 import json
 import os
 
-from example.commons import Collector
 from pyecharts import options as opts
 from pyecharts.charts import Page, TreeMap
+from pyecharts.faker import Collector
 
 C = Collector()
