+++ test/test_geo.py
@@ -2,9 +2,9 @@ from unittest.mock import patch
 
 from nose.tools import assert_equal, assert_in
 
-from example.commons import Faker
 from pyecharts import options as opts
 from pyecharts.charts import Geo
+from pyecharts.faker import Faker
 
 
 @patch("pyecharts.render.engine.write_utf8_html_file")
