+++ test/test_map3d.py
@@ -2,8 +2,8 @@ from unittest.mock import patch
 
 from nose.tools import assert_equal, assert_in
 
-from example.commons import Faker
 from pyecharts.charts import Map3D
+from pyecharts.faker import Faker
 
 
 @patch("pyecharts.render.engine.write_utf8_html_file")
