+++ test/test_page.py
@@ -9,10 +9,10 @@ from nose.tools import (
     raises,
 )
 
-from example.commons import Faker
 from pyecharts.charts import Bar, Line, Page
 from pyecharts.commons.utils import OrderedSet
 from pyecharts.components import Table
+from pyecharts.faker import Faker
 
 
 def _create_bar() -> Bar:
